Added functionality to change between views with the nav bar.
This is done by setupMiddleSection.js, which is a class that is called whenever the user clicks a nav_bar button.

PLus of course... I added and changed a littel more styling and tried playing around. The chat windows will not have a scroll bar when I put it in the bottom like I have done apparently, and when I change it it looks weird, this will do for now.